Add support for filtering tests using a regular expression
Golang supports filtering tests using a -run flag, which can be a
regular expression. The e2e tests don't honor this flag since we're
building the tests in a particular way so we can run parallel tests
(e.g., non-destructive) tests before running the serial tests, which
test remediations and other actions that affect the state of the
cluster, potentially affecting other tests.

To make things more egonomic for contributors, it would be nice to run a
single test so you don't have to wait for the entire suite to execute.

This commit adds a framework flag that introduces a new command line
argument called `-testRegex` and filters e2e tests using the flag as if
it were a regular expression.

Fixes #226