OCPBUGS-9382: Expose validation issue when creating TailoredProfiles
TailoredProfiles can extend existing profiles, which can be either Node
or Platform type.

However, it's possible to create a TailoredProfile that extends a
profile of a patricular type, and then reference rules of the opposite
type. This causes issues during scans because you'd expect the rules to
be excluded, but they're not.

This commit adds an e2e test the exposes the issue. This can be
addressed with improved validation when creating a TailoredProfile.

Related to issue #65.