Implement an OCI image for osde2e testing
To test the Compliance Operator on different platforms, like Red Hat
OpenShift on AWS (ROSA), we can use the openshift/osde2e framework.

To do this, we need to write tests in a specific format, and build them
into a container image that can be invoke on a managed platform.

This commit introduces a rough idea of what that could look like.