CMP-3077: Update kustomize manifest to use Konflux image references
We have Konflux building images for each component, and pushing them to
a public registry in Quay. This commit updates manifests to use those
image references. This will allows us to consolidate some of our tooling
so we're not building the same image and pushing it to different places.

A subsequent patch will enable a Konflux feature to automatically keep
these image references up-to-date.