CMP-3077: Implement bundle builds for redhat registry
Implement a containerfile that builds a bundle image with references to
images from registry.redhat.io. This allows Konflux to build a shippable
bundle image that will include references to the images available from
registry.redhat.io.

This will be available alongside a bundle that still contains quay.io
references for upstream deployments.