Use make to set versions in manifests
In order to generate proper manifests, we need the previous version of
the operator, and the current version. This commit seds the CSV manifest
so that it includes the current version, and the most recent released
version.

This will make it easier to generate bundle images with valid version
information, which will be reflected when users install the operator,
along with having a valid upgrade path.

This approach is also baked into the bundle target, meaning CI will fail
if there is any drift between the version and the generated manifests,
making it more likely to keep generated manifests up-to-date with the
version we're planning to release.