CMP-3121: Increase unit test timeout to 20 minutes
We're seeing arm64 builds timeout because 10 minutes isn't enough time
to run the unit tests. However, when testing this on a native arm64
machine, the run time difference is negligible. One possibility is that
because the Github Action runner is using amd64 under the hood, it's
taking longer to cross-compile and build for arm64.

This commit doubles the unit test timeout from 10m to 20m to see if this
helps improve arm64 build stability.