CMP-3156: Simplify similar cases in ComplianceScan pending phase
We need to remove annotations when rerunning a scan, regardless of it
being rerun for a timeout or a rescan. We handled these cases
separately, but didn't need to since we're using `delete()`, which
performs a no-op if the annotation doesn't exist.

This PR simplifies the code by reducing both conditionals into a single
case.