OCPBUGS-55181: Update NERC-CIP reference
The NERC-CIP references in the ComplianceAsCode/content were updated
recently because they were pointing to a dead link (HTTP 404).

  https://github.com/ComplianceAsCode/content/pull/12892

This commit updates the profile parser to use the same link. Without
doing this, the rules won't contain annotations or controls for the
NERC-CIP profile, even though the Compliance Operator has a profile for
it.