CMP-3515: Use Konflux builds for content
Konflux already builds the compliance operator content and pushes it to
Quay. This means we don't need to setup duplicate Github Actions to
maintain the content and push it to k8scontent:latest.

This commit updates the project by replacing references to
k8scontent:latest with the corresponding references in Konflux.