CMP-3728: Move platform helper functions to utils
The operator package has a bunch of useful utility methods for detecting
the platform, and returning the default roles and auto-remediation
feature support for those platforms. We should be using these pieces of
information in other parts of the operator, like scan setting validation
where we emit a warning/error if the user is attempting to create a scan
setting with auto-remediation enabled on a platform that doesn't use it.

This commit moves those helper functions to the utility package where we
can reuse them.