CMP-3895: Improve rescan wait logic for e2e tests
The e2e tests have been failing with what appears to be a transient race
condition when a test rescans the environment, then waits for the suite
or scan to get to a terminal state, like DONE and NON-COMPLIANT.

The problem is that the test does something to initiate a rescan
(deleting a suite, creating a scan setting binding, etc), and then it
immediately starts looking for the ComplianceSuite status to see if it's
done.

In the case of rescans, the data from the last scan might still be
lingering around and if the test is faster than the reconcile loops of
the ComplianceScan controller, it could breeze through the wait logic
and think the scan is already complete.

This commit attempts to address that by using the rescan annotation. By
using the annotation to rescan (instead of deleting a scan suite
entirely), we can update the function that polls the suite status to
also check for the annotations. If the wait function sees the
annotation, it should continue polling since that's a clear signal that
the ComplianceScan controller has not updated the scan data or reset the
states back to PENDING. If none of the scans in the suite have the
annotation, then it's probably safe to move forward and assume the
status for each scan is updated and accurate, since the ComplianceScan
controller will remove the rescan annotation when it restarts the scan.