CMP-3765: Bootstrap samples for OpenShift Virt profile (#813)
Let's use the config samples to show how users can write CustomRule
resources to harden an OpenShift virtualization cluster. These updates
are isolated to an openshift-virtualization directory with a
kustomization file so that they're easy to reuse and apply in existing
clusters (e.g., oc apply -k config/samples/custom-rules/openshift-virtualization/).

This idea can be expanded on to show how other contributors and teams
can ship their own content for the Compliance Operator, and bundle it
such that it can be reused by others.

One potential future workflow would be reuse this idea by putting
necessary rules in a directory, build the rules and rbac changes into a
scratch image, upload the image to registry, and reuse the profile
bundle feature to parse the rules and load them into the compliance
operator.