CMP-3908: Remove auth header from metrics endpoint test
The TestResultsServerHTTPVersion test keeps failing with an error about
not being able to run the curl command to query the metrics endpoint.
The command tries to fish the sa token from disk, but it's returning an
error code, which is tripping up the test (likely because the pod hasn't
mounted the service account token?). Trying it without the auth header
still returns the metrics, and simplifies the command.