CMP-4008: Eliminate race-condition in TestScheduledSuite
We're noticing issues in TestScheduledSuite where the test fails because
an update the to ComplianceSuite fails. We suspect this is due to the
copy of the suite getting updated by the controller before the update is
applied.

This commit swaps the `Update` for a `Patch`, which doesn't require the
whole ComplianceSuite to execute the update. Instead it just needs the
thing we're changing (e.g., the schedule), and applies that to the
latest version of the resource.